% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{add_modules_to_phybreak}
\alias{add_modules_to_phybreak}
\title{Create a phybreak-object from data and prior distributions.}
\usage{
add_modules_to_phybreak(
  le,
  multiple.introductions = TRUE,
  spatial = FALSE,
  contact = FALSE,
  infectivity = FALSE,
  phyb.obj = TRUE,
  ...
)
}
\description{
Create a phybreak-object from data and prior distributions.
}
