% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{AvianFluH7N7_2003}
\alias{AvianFluH7N7_2003}
\title{Avian influenza (H7N7) epidemic}
\format{A list with four elements:
\describe{
  \item{sequences_HA}{the HA gene}
  \item{sequences_NA}{the NA gene}
  \item{sequences_PB2}{the PB2 gene}
  \item{sampledays}{the days at which the samples were taken (or the detection day + 2 days, for the 10 
  unsampled farms)}
}}
\usage{
AvianFluH7N7_2003
}
\description{
DNA sequences (base order randomized) of NA, HA, and PB2 genes of samples taken from farms during the Dutch avian
  influenza (H7N7) epidemic in 2003. The dataset contains 241 farms, of which 231 were sampled and sequenced 
  (for the other 10, identified by the prefix UNK, only the detection day + 2 days is given, as in Klinkenberg 
  et al (2017)). The labels of the sequences refer to the labels in GISAID (\url{www.gisaid.org}), where the 
  sequences in correct base order can be found.
}
\references{
\href{http://dx.doi.org/10.1371/journal.ppat.1002094}{Bataille et al. (2011)} Evolutionary 
  analysis of inter-farm transmission dynamics in a highly pathogenic avian influenza epidemic. 
  \emph{PLoS Pathog}, \strong{7}(6): e1002094. 
    
  \href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
    inference of phylogenetic and transmission trees in infectious disease outbreaks. 
    \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\keyword{datasets}
