% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_phybreak.R
\name{c.phybreak}
\alias{c.phybreak}
\title{Combine mcmc-chains of phybreak-objects.}
\usage{
\method{c}{phybreak}(...)
}
\arguments{
\item{...}{Objects of class \code{phybreak} to be concatenated.}
}
\value{
The first \code{phybreak}-object with mcmc samples of all objects.
}
\description{
The function combines the sample slots of its arguments. The variable slot (current state) is taken from the first argument
}
\examples{
#First create a phybreak-object
simulation <- sim_phybreak(obsize = 5)
MCMCstate1 <- phybreak(dataset = simulation)
MCMCstate1 <- burnin_phybreak(MCMCstate1, ncycles = 20)
MCMCstate1 <- sample_phybreak(MCMCstate1, nsample = 50, thin = 2)

MCMCstate2 <- phybreak(dataset = simulation)
MCMCstate2 <- burnin_phybreak(MCMCstate2, ncycles = 20)
MCMCstate2 <- sample_phybreak(MCMCstate2, nsample = 50, thin = 2)

MCMCstate <- c(MCMCstate1, MCMCstate2)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
