% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveSize_phybreak.R
\name{ESS}
\alias{ESS}
\alias{effectiveSize_phybreak}
\alias{effectiveSize_factor}
\title{Effective sample size}
\usage{
effectiveSize_phybreak(x)

effectiveSize_factor(chain)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{chain}{Vector of class \code{factor}, or a numeric vector with integers indicating categories.}
}
\value{
Effective sample sizes.
}
\description{
Effective sample size
}
\details{
For all parameters and continuous variables (infection times), \code{effectiveSize_phybreak} calculates
  the effective sample size (ESS) with the \code{\link[coda]{effectiveSize}} function in \pkg{coda}. For the infectors, 
  a method is used that is similar to the method for the approximate ESS for phylogenetic trees, described in 
  Lanfaer et al (2016):
  \enumerate{
    \item Define as distance measure between two sampled infectors \eqn{D(i,j) = 0} if \eqn{i = j}, 
        and \eqn{D(i,j) = 1} if \eqn{i \neq= j}{i <> j}
    \item Calculate the mean squared distance f(k) between sampled infectors at intervals k = 1,2,... in the mcmc chain.
        The distance will increase with increasing interval k.
    \item Use the rate at which f(k) approaches the asymptote to calculate the ESS (see Lanfaer et al, 2016)
  }
  This method can also be directly called with \code{effectiveSize_factor} for a single vector of categorical variables.
}
\section{Functions}{
\itemize{
\item \code{effectiveSize_phybreak}: Effective sample size of phybreak posterior.

\item \code{effectiveSize_factor}: Effective sample size of a categorical variable.
}}

\examples{
#First create a phybreak object
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(data = simulation)

MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 50, thin = 2)
effectiveSize_phybreak(MCMCstate)
}
\references{
\href{http://dx.doi.org/10.1093/gbe/evw171}{Lanfaer et al. (2016)} Estimating 
  the effective sample size of tree topologies from Bayesian phylogenetic analyses. 
  \emph{Genome Biol Evol}, \strong{8}(8): 2319-2332.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
