% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FMD_2001}
\alias{FMD_2001}
\title{Foot-and-Mouth Disease outbreak (2001)}
\format{A matrix with 15 rows (one per farm) and 8196 column (one per nucleotide); the names contain sampling days}
\usage{
FMD_2001
}
\description{
DNA nucleotide patterns and sampling days of FMD virus from 15 farms in the UK. The nucleotides per sample 
  are not ordered as a virus sequence, but do contain the correct numbers of each nucleotide.
}
\references{
\href{http://dx.doi.org/10.1098/rspb.2007.1442}{Cottam et al. (2008)}
  Integrating genetic and epidemiological data to determine transmission pathways 
  of foot-and-mouth disease virus. \emph{Proc R Soc B}, \strong{275}: 887-895.
}
\keyword{datasets}
