% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FMD_2007}
\alias{FMD_2007}
\title{Foot-and-Mouth Disease outbreak (2007)}
\format{A list with two elements:
\describe{
  \item{sequences}{the sequences (class \code{DNAbin})}
  \item{dates}{the sampling days (class \code{Date})}
}}
\usage{
FMD_2007
}
\description{
DNA sequences and sampling days of FMD virus from 10 farms in the UK (11 sequences)
}
\references{
\href{http://dx.doi.org/10.1371/journal.ppat.1000050}{Cottam et al. (2008)} Transmission pathways 
  of foot-and-mouth disease virus in the United Kingdom in 2007. \emph{PLoS Pathog}, \strong{4}(4): e1000050.
}
\keyword{datasets}
