% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phybreak.R
\name{get_phybreak}
\alias{get_phybreak}
\alias{get_data}
\alias{get_parameters}
\alias{get_transtree}
\alias{get_phylo}
\alias{get_multiPhylo}
\alias{get_mcmc}
\alias{get_bottlenecks}
\title{Extracting from a phybreak object}
\usage{
get_data(x)

get_parameters(x, whichpars = "posterior")

get_transtree(x)

get_phylo(x, simmap = FALSE, samplenr = 0)

get_multiPhylo(x, thin = 1, nkeep = Inf)

get_mcmc(x, thin = 1, nkeep = Inf)

get_bottlenecks(x)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{whichpars}{Which parameters to return. Either a vector with parameter names, or \code{"all"} for all parameters, or 
\code{"posterior"} for parameters for which a posterior is sampled.}

\item{simmap}{Whether to include class \code{"simmap"} elements (package \pkg{phytools}), colouring the branches
on the tree to indicate hosts. Is used by \code{\link{plotPhylo}}.}

\item{samplenr}{The posterior tree sample to choose. If \code{samplenr = 0}, the current state is used.}

\item{thin}{Thinning interval.}

\item{nkeep}{Number of samples to keep, counting from tail of the chain.}
}
\description{
Extracting from a phybreak object
}
\section{Functions}{
\itemize{
\item \code{get_data}: The data in class \code{phybreakdata}.

\item \code{get_parameters}: A named vector with current parameter values.

\item \code{get_transtree}: A \code{data.frame} with current infectors and infection times.

\item \code{get_phylo}: Returns an object of class \code{\link[ape]{phylo}} ans optionally of class
\code{"simmap"} (package \pkg{phytools}).

\item \code{get_multiPhylo}: Returns an object of class \code{\link[ape]{multiphylo}}.

\item \code{get_mcmc}: An object of class \code{"mcmc"} (package \pkg{coda}), with sampled parameters, 
infection times, and infectors.

\item \code{get_bottlenecks}: A vector with for each host the current number of extant lineages in its phylogenetic minitree
at the time of infection (the bottleneck size).
}}

\examples{
#First build a phybreak-object.
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(dataset = simulation)
MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 50, thin = 2)

get_transtree(MCMCstate)
get_parameters(MCMCstate)
mcmcobject <- get_mcmc(MCMCstate, thin = 2)
plot.phylo(get_phylo(MCMCstate))
get_data(MCMCstate)
get_bottlenecks(MCMCstate)

#function from package phangorn:
phangorn::parsimony(get_phylo(MCMCstate), get_data(MCMCstate)$sequences)

tree0 <- get_phylo(MCMCstate)
seqdata <- get_data(MCMCstate)$sequences
phangorn::pml(tree0, seqdata, 
              rate = 0.75*get_parameters(MCMCstate, "mu")) 
logLik(MCMCstate, genetic = TRUE, withinhost = FALSE, 
       sampling = FALSE, generation = FALSE) #should give the same result as 'pml'
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
