% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect_distribution.R
\name{infect_distribution}
\alias{infect_distribution}
\title{Infectiousness distribution.}
\usage{
infect_distribution(time, vars, p)
}
\arguments{
\item{time}{The proposed infection time of the current host}

\item{vars}{The list of variables describing the nodes in the tree. This list should 
at least include the infection times of the hosts. For \code{p$trans.model = "sample+culling"}
also nodetimes of the sample nodes and the culling times of the hosts are needed as input.}

\item{p}{The list with parameters for the model.}
}
\value{
The transmission probabilities of all hosts infected before the current host. For each host
  with an infection time before the proposed infection time, the probability of being an infector
  of the current host is calculated according to the given infectiousness function.
}
\description{
This function calculates the probability of transmission given the infectiousness 
distribution for the hosts.
}
