% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{M_tuberculosis_2013}
\alias{M_tuberculosis_2013}
\title{Mycobacterium tuberculosis outbreak}
\format{A list with two elements:
\describe{
  \item{sequences_Mtb}{the sequences (class \code{DNAbin})}
  \item{dates_Mtb}{the sampling days (class \code{Date})}
}}
\usage{
M_tuberculosis_2013
}
\description{
DNA SNP patterns and sampling days of an M. tuberculosis outbreak with 33 cases in Canada (Didelot et al, 2013). 
  The dataset contained only 20 SNP patterns with unspecified nucleotides, so the sequences in this dataset
   contain only `a` and `c` on these 20 loci. Additional loci with only `a` are added to a total sequence length
   of 440,000, which is 10% of the actual M. tuberculosis genome.
}
\references{
\href{http://dx.doi.org/10.1093/molbev/msu121}{Didelot et al. (2013)} Bayesian inference 
  of infectious disease transmission from whole-genome sequence data. \emph{Mol Biol Evol}, \strong{31}(7): 1869-1879.
}
\keyword{datasets}
