% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree.R
\name{phylotree}
\alias{phylotree}
\title{Maximum clade credibility tree.}
\usage{
phylotree(phybreak.object, samplesize = Inf, support = c("proportion",
  "count"), clade.times = TRUE, time.quantiles = c(0.025, 0.5, 0.975),
  phylo.class = FALSE)
}
\arguments{
\item{phybreak.object}{An object of class \code{phybreak}.}

\item{samplesize}{The number of posterior samples that is used, taken from the tail.}

\item{support}{Whether to return the support (= posterior probability) for each infector as a \code{"proportion"} 
or as a \code{"count"} of posterior trees in which that transmission link or transmission cluster is present.}

\item{clade.times}{Whether to include summary times for root nodes of clades, based only on posterior trees 
with the same clade as in the consensus tree. These are used to calculate a mean and standard deviation.}

\item{time.quantiles}{Used only if \code{clade.times = TRUE}.}

\item{phylo.class}{Whether to return an object of class \code{"phylo"}, in which case a single tree from the 
posterior is returned (not with summary infection times).}
}
\value{
A \code{data.frame} with per item (=node) its parent and support per clade, and optionally summary node times. 
  The first n nodes are the samples, the last n-1 nodes are the internal nodes.
}
\description{
Identify the maximum clade credibility tree from a \code{phybreak}-object containing posterior samples.
}
\examples{
#First build a phybreak-object containing samples.
simulation <- sim.phybreak(obsize = 20)
MCMCstate <- phybreak(simulation)
MCMCstate <- burnin.phybreak(MCMCstate, ncycles = 200)
MCMCstate <- sample.phybreak(MCMCstate, nsample = 100, thin = 10)

phylotree(MCMCstate)
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}

