% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_phybreak.R
\name{print.phybreak}
\alias{print.phybreak}
\title{Description of phybreak object.}
\usage{
\method{print}{phybreak}(x, ...)
}
\arguments{
\item{x}{A \code{phybreak} object.}

\item{...}{Some methods for this generic require additional arguments. None are used in this method.}

\item{printlen}{The number of host names to print.}
}
\description{
Give a short description of the data and the model, with the current parameter values.
}
\examples{
simulation <- sim_phybreak()
x <- phybreak(simulation)
print(x)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
