% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_phybreak.R
\name{quantile.phybreak}
\alias{quantile.phybreak}
\title{Quantiles of sampled parameter values.}
\usage{
\method{quantile}{phybreak}(
  x,
  whichpars = "posterior",
  probs = c(0.025, 0.5, 0.975),
  type = 7,
  ...
)
}
\arguments{
\item{whichpars}{Which parameters to return. Either a vector with parameter names, or \code{"all"} for all parameters, or 
\code{"posterior"} for parameters for which a posterior is sampled.}

\item{probs}{Numeric vector of probabilities with values in \eqn{[0,1]}.}

\item{type}{An integer between 1 and 9 selecting one of the nine quantile algorithms (see \code{\link[stats]{quantile}}).}
}
\value{
A named matrix with desired quantiles. For parameters that are not estimated, all quantiles give the fixed value.
}
\description{
Quantiles of sampled parameter values.
}
\examples{
#First build a phybreak-object.
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(data = simulation)
MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 100)

quantile(MCMCstate)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
