% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phybreak_parallel.R
\name{sample_phybreak_parallel}
\alias{sample_phybreak_parallel}
\title{Sampling from a phybreak MCMC-chain.}
\usage{
sample_phybreak_parallel(
  x,
  nsample,
  thin = 1,
  thinswap = 1,
  classic = 0,
  keepphylo = 0,
  withinhost_only = 0,
  parameter_frequency = 1,
  status_interval = 10,
  histtime = -1e+05,
  history = FALSE,
  nchains = 1,
  heats = NULL,
  all_chains = FALSE,
  outfile = ""
)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{nsample}{The number of samples to take.}

\item{thin}{The thinning to use (values after every \code{thin}'th iteration will be included in the posterior). 
Each iteration does one update of all parameters and tree updates with each host as focal host once.}

\item{classic}{The proportion of tree updates with the classic protocol (reference see below), in which within-host
minitrees are proposed by simulating coalescent times and tree topology. In the current default protocol only 
coalescent times are proposed with the minitree topology kept intact. This is followed by removing and reconnecting
the sampling tips one by one. This results in better mixing of the mcmc-chain if there is much 
genetic information (many SNPs) and/or if there are many possible within-host minitree topologies 
(e.g. many samples per host). The classic protocol is faster in terms of updates/second and can thus be more efficient
with little genetic information.}

\item{keepphylo}{The proportion of tree updates keeping the phylogenetic tree intact, only possible if there is one 
sample per host and the \code{wh.model = "linear"} with complete bottleneck.}

\item{withinhost_only}{The proportion of tree updates in which only the within-host minitree is sampled, and 
the transmission tree and infection times are kept unchanged.}

\item{parameter_frequency}{The relative frequency by which the model parameters are updated relative to updating each host.}

\item{status_interval}{The number of seconds between each on-screen print of the progress of the mcmc-chain.}
}
\value{
The \code{phybreak} object used to call the function, including (additional) samples from the posterior.
}
\description{
Function to take (additional) samples from the posterior distribution of a phylogenetic and transmission tree 
  (plus associated parameters), within a \code{phybreak} object (\code{sample.phybreak} is deprecated).
}
\examples{
#First create a phybreak-object
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(dataset = simulation)

MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 50, thin = 2)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
