% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin-phybreak.R
\name{burnin_phybreak}
\alias{burnin_phybreak}
\alias{burnin.phybreak}
\title{MCMC updating of a phybreak-object.}
\usage{
burnin_phybreak(x, ncycles, keepphylo = NULL, withinhost_only = 0)

burnin.phybreak(...)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{ncycles}{Number of iterations to be carried out. Each iteration does one update of all parameters and
tree updates with each host as focal host once.}

\item{keepphylo}{The proportion of tree updates keeping the phylotree intact, only possible if there is one sample per host
and the \code{wh.model = "linear"}. If \code{NULL} (default), it is set to 0.2 only in that case, otherwise to 0.}

\item{withinhost_only}{The proportion of tree updates in which only the within-host minitree is sampled, and 
the transmission tree and infection times are kept unchanged.}
}
\value{
The \code{phybreak} object provided as input, with variables and parameters changed due to the updating.
}
\description{
This function allows the MCMC chain to burn in. If used after samples have been taken (with \code{\link{sample_phybreak}}), 
  these samples will be returned unchanged in the output (\code{burnin.phybreak} is deprecated).
}
\examples{
#First create a phybreak object
simulation <- sim.phybreak(obsize = 5)
MCMCstate <- phybreak(data = simulation)

MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 50)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
