% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{contact_parameters}
\alias{contact_parameters}
\title{Module for contact data}
\usage{
contact_parameters(
  le,
  cnt.invest.trans = 1,
  cnt.invest.nontrans = 0.5,
  cnt.report = 1,
  cnt.report.false = 0,
  cnt.types = c(0.5, 0.5),
  est.cnt.types = F,
  est.cnt.invest.trans = T,
  est.cnt.invest.nontrans = T,
  est.cnt.report = T,
  prior.cnt.eta.alpha = 1,
  prior.cnt.eta.beta = 1,
  prior.cnt.lambda.alpha = 1,
  prior.cnt.lambda.beta = 1,
  prior.cnt.epsilon.alpha = 1,
  prior.cnt.epsilon.beta = 1
)
}
\arguments{
\item{cnt.invest.trans}{Probability for transmission pair that contact took place (\eqn{\eta})}

\item{cnt.invest.nontrans}{Probability for non-transmission pair that contact took place (\eqn{\lambda})}

\item{cnt.report}{Reporting coverage - probability that contact was reported (\eqn{\epsilon})}

\item{cnt.report.false}{False reporting probability - probability that report was false (\eqn{\zeta})}

\item{est.cnt.invest.trans}{Estimate \eqn{\eta}, default = TRUE}

\item{est.cnt.invest.nontrans}{Estimate \eqn{\lambda}, default = TRUE}

\item{est.cnt.report}{Estimate \eqn{\epsilon}, default = TRUE}

\item{prior.cnt.eta.alpha}{Beta distributed prior for \eqn{\eta}, \eqn{\alpha} parameter}

\item{prior.cnt.eta.beta}{Beta distributed prior for \eqn{\eta}, \eqn{\beta} parameter}

\item{prior.cnt.lambda.alpha}{Beta distributed prior for \eqn{\lambda}, \eqn{\alpha} parameter}

\item{prior.cnt.lambda.beta}{Beta distributed prior for \eqn{\lambda}, \eqn{\beta} parameter}

\item{prior.cnt.epsilon.alpha}{Beta distributed prior for \eqn{\epsilon}, \eqn{\alpha} parameter}

\item{prior.cnt.epsilon.beta}{Beta distributed prior for \eqn{\epsilon}, \eqn{\beta} parameter}
}
\description{
Module for contact data
}
