% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak.R
\name{phybreak}
\alias{phybreak}
\title{Create a phybreak-object from data and prior distributions.}
\usage{
phybreak(
  dataset,
  times = NULL,
  introductions = 1,
  mu = NULL,
  gen.shape = 3,
  gen.mean = 1,
  trans.model = "gamma",
  trans.init = 1e-04,
  trans.growth = 1,
  trans.sample = 0.1,
  trans.culling = 5,
  sample.shape = 3,
  sample.mean = 1,
  wh.model = "linear",
  wh.bottleneck = "auto",
  wh.history = 1,
  wh.slope = 1,
  wh.exponent = 1,
  wh.level = 0.1,
  dist.model = "power",
  dist.exponent = 2,
  dist.scale = 1,
  dist.mean = 1,
  est.gen.mean = TRUE,
  prior.gen.mean.mean = 1,
  prior.gen.mean.sd = Inf,
  est.sample.mean = TRUE,
  prior.sample.mean.mean = 1,
  prior.sample.mean.sd = Inf,
  est.trans.growth = TRUE,
  est.trans.sample = TRUE,
  est.wh.history = TRUE,
  est.wh.slope = TRUE,
  prior.wh.slope.shape = 3,
  prior.wh.slope.mean = 1,
  est.wh.exponent = TRUE,
  prior.wh.exponent.shape = 1,
  prior.wh.exponent.mean = 1,
  est.wh.level = TRUE,
  prior.wh.level.shape = 1,
  prior.wh.level.mean = 0.1,
  est.dist.exponent = TRUE,
  prior.dist.exponent.shape = 1,
  prior.dist.exponent.mean = 1,
  est.dist.scale = TRUE,
  prior.dist.scale.shape = 1,
  prior.dist.scale.mean = 1,
  est.dist.mean = TRUE,
  prior.dist.mean.shape = 1,
  prior.dist.mean.mean = 1,
  use.tree = FALSE,
  ...
)
}
\arguments{
\item{dataset}{An object with sequences plus additional data (class \code{'phybreakdata'}). 
Data contain \code{sequences} and \code{sampling.times}, and potentially \code{sim.infection.times}, 
\code{sim.infectors}, and \code{sim.tree}. Prepare your data in this format
by \code{\link{phybreakdata}} or by simulation with \code{\link{sim.phybreak}}.

It is also possible to provide only sequences as data, (class \code{'DNAbin'}, \code{'phyDat'}, or a \code{matrix} with nucleotides, 
each row a host, each column a nucleotide), and corresponding sampling times in the separate \code{times} argument.}

\item{times}{Vector of sampling times, needed if the data consist of only sequences. If the vector is named,
these names will be used to identify the hosts.}

\item{mu}{Initial value for mutation rate (defined per site per unit of time). If \code{NULL} (default), then an initial
value is calculated by dividing the number of SNPs by the product: 0.75 times 'total sequence length' times 'sum of
edge lengths in the initial phylogenetic tree'.
NOTE: mutation is defined as assignment of a random nucleotide at a particular site; this could be the 
nucleotide that was there before the mutation event. Therefore, the actual rate of change of nucleotides 
is \code{0.75*mu}.}

\item{gen.shape}{Shape parameter of the generation interval distribution (not estimated).}

\item{gen.mean}{Initial value for the mean generation interval, i.e. the interval between infection of a secondary
case by a primary case.}

\item{sample.shape}{Shape parameter of the sampling interval distribution (not estimated), i.e. the interval between
infection and sampling of a host.}

\item{sample.mean}{Initial value for the mean sampling interval.}

\item{wh.model}{The model for within-host pathogen dynamics (effective pathogen population size = 
N*gE = actual population size * pathogen generation time), used to simulate coalescence events. Names and numbers are allowed.
Options are:
\enumerate{
  \item "single": effective size = 0, so coalescence occurs 'just before' transmission in the infector (complete bottleneck)
  \item "infinite": effective size = Inf, with complete bottleneck, so coalescence occurs 'just after' transmission in the infectee
  \item "linear": effective size at time t after infection = \code{wh.level + wh.slope * t} (complete or wide bottleneck; if complete, \code{wh.level = 0})
  \item "exponential": effective size at time t after infection = \code{wh.level * exp(wh.exponent * t)} (wide bottleneck)
  \item "constant": effective size = wh.level (wide bottleneck)
}}

\item{wh.bottleneck}{Whether the bottleneck should be complete or wide, which is only an option if \code{wh.model = "linear"} 
(in that case, \code{"auto"} defaults to \code{"complete"}).}

\item{wh.slope}{Initial value for the within-host slope, used if \code{wh.model = "linear"}.}

\item{wh.exponent}{Initial value for the within-host exponent, used if \code{wh.model = "exponential"}}

\item{wh.level}{Initial value for the within-host effective pathogen size at transmission, used if \code{wh.bottleneck = "wide"}
(if \code{wh.model = "exponential"} or \code{"constant"}, and optional if \code{wh.model = "linear"})}

\item{est.gen.mean}{Whether to estimate the mean generation interval or keep it fixed.}

\item{prior.gen.mean.mean}{Mean of the (gamma) prior distribution of mean generation interval \code{mG} 
(only if \code{est.gen.mean = TRUE}).}

\item{prior.gen.mean.sd}{Standard deviation of the (gamma) prior distribution of mean generation interval \code{mG} 
(only if \code{est.gen.mean = TRUE}).}

\item{est.sample.mean}{Whether to estimate the mean sampling interval or keep it fixed.}

\item{prior.sample.mean.mean}{Mean of the (gamma) prior distribution of mean sampling interval \code{mS} 
(only if \code{est.sample.mean = TRUE}).}

\item{prior.sample.mean.sd}{Standard deviation of the (gamma) prior distribution of mean sampling interval \code{mS} 
(only if \code{est.sample.mean = TRUE}).}

\item{est.wh.slope}{Whether to estimate the within-host slope or keep it fixed.}

\item{prior.wh.slope.shape}{Shape parameter of the (gamma) prior distribution of \code{wh.slope} 
(only if \code{est.wh.slope = TRUE}).}

\item{prior.wh.slope.mean}{Mean of the (gamma) prior distribution of \code{wh.slope} 
(only if \code{est.wh.slope = TRUE}).}

\item{est.wh.exponent}{Whether to estimate the within-host exponent or keep it fixed.}

\item{prior.wh.exponent.shape}{Shape parameter of the (gamma) prior distribution of \code{wh.exponent} 
(only if \code{est.wh.exponent = TRUE}).}

\item{prior.wh.exponent.mean}{Mean of the (gamma) prior distribution of \code{wh.exponent} 
(only if \code{est.wh.exponent = TRUE}).}

\item{est.wh.level}{Whether to estimate the within-host level at \code{t = 0} or keep it fixed.}

\item{prior.wh.level.shape}{Shape parameter of the (gamma) prior distribution of \code{wh.level} 
(only if \code{est.wh.level = TRUE}).}

\item{prior.wh.level.mean}{Mean of the (gamma) prior distribution of \code{wh.level} 
(only if \code{est.wh.level = TRUE}).}

\item{use.tree}{Whether to use the transmission and phylogenetic tree given in data of class \code{'obkData'}, 
to create a \code{phybreak}-object with an exact copy of the outbreak. This requires more data in \code{data}: 
the slot \code{individuals} with vectors \code{infector} and \code{date}, and the slot \code{trees} with at least 
one phylogenetic tree. Such data can be simulated with \code{\link{sim.phybreak}}.}

\item{...}{If arguments from previous versions of this function are used, they may be interpreted correctly through 
this argument, but it is better to provide the correct argument names.}
}
\value{
An object of class \code{phybreak} with the following elements
  \describe{
    \item{d}{a \code{list} with data, i.e. names, sequences, sampling times, and total number of SNPs.}
    \item{v}{a \code{list} with current state of all nodes in the tree: times, hosts in which they reside,
      parent nodes, node types (sampling, coalescent, or transmission)}
    \item{p}{a \code{list} with the parameter values}
    \item{h}{a \code{list} with helper information for the MCMC-method: \code{si.mu} and \code{si.wh} for 
      efficiently proposing \code{mu} and \code{slope}, matrix \code{dist} with weights for infector sampling 
      based on sequence distances, \code{logical}s \code{est.mG}, \code{est.mS}, and \code{est.wh.slope} whether to estimate 
      mean generation interval \code{mG}, mean sampling interval \code{mS}, and within-host \code{slope}, 
      and parameters for the priors of \code{mG}, \code{mS}, and \code{slope}.
    }
    \item{s}{an empty \code{list} that will contain vector and matrices with the posterior samples; 
      in matrices, the rows are nodes in the phylogenetic tree, the columns are the samples}
  }
}
\description{
phybreak takes as data a \code{\link{phybreakdata}}-object with sequences (individuals in rows, 
nucleotides in columns) and sampling times, and potentially more. Parameter values are used as 
initial values in the MCMC-chain or kept fixed. All variables are initialized by random samples 
from the prior distribution, unless a complete tree is given in the data and should be used 
(\code{use.tree = TRUE}). It is also possible to provide only sequences as data, and sampling times separately.
}
\examples{
simulation <- sim_phybreak(obsize = 10)
MCMCstate <- phybreak(dataset = simulation)

simulation <- sim_phybreak(obsize = 10)
MCMCstate <- phybreak(dataset = simulation, use.tree = TRUE)


sampletimedata <- c(0,2,2,4,4)
sampleSNPdata <- matrix(c("a","a","a","a","a",
                          "a","c","c","c","c",
                          "t","t","t","g","g"), nrow = 5)
dataset <- phybreakdata(sequences = sampleSNPdata, sample.times = sampletimedata)
MCMCstate <- phybreak(data = dataset)

### also possible without 'phybreakdata' as intermediate, 
### but only with single samples per host and not with additional data (future implementation)
MCMCstate <- phybreak(data = sampleSNPdata, times = sampletimedata)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
