% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreakdata.R
\name{phybreakdata}
\alias{phybreakdata}
\title{Create a phybreakdata-object from raw data.}
\usage{
phybreakdata(sequences, sample.times, host.names = NULL,
  sim.infection.times = NULL, sim.infectors = NULL, sim.tree = NULL,
  sim.reassortment = NULL)
}
\arguments{
\item{sequences}{Sequence data of class \code{'DNAbin'}, \code{'phyDat'}, or a \code{matrix} with nucleotides, 
each row a host, each column a nucleotide). All nucleotides that are not \code{'a'}, \code{'c'}, \code{'g'}, 
or \code{'t'}, will be turned into \code{'n'}.}

\item{sample.times}{A vector of sampling times (\code{numerical} or \code{Date}).}

\item{host.names}{A vector with host names.}

\item{sim.infection.times}{A vector with infection times (\code{numerical} or \code{Date}).}

\item{sim.infectors}{A vector with infectors, either by name or by position (use 0 for the index case).}

\item{sim.tree}{A tree of class \code{'phylo'}, with tip names identifying the hosts.}
}
\value{
An object of class \code{phybreakdata} with the following elements
  \describe{
    \item{sequences}{a \code{'phyDat'}-object with the sequence data.}
    \item{sample.times}{a named \code{vector} with the sample times.}
    \item{sim.infection.times}{a named \code{vector} with the (simulated) infection times (if provided).}
    \item{sim.infectors}{a named \code{vector} with the (simulated) infectors (if provided).}
    \item{sim.tree}{a \code{'phylo'}-object with the (simulated) phylogenetic tree (if provided).}
  }
}
\description{
phybreakdata takes as data sequences and sampling times and makes a \code{phybreakdata} object.
Each sample is assumed to be associated with a separate host. The number of sequences
should be equal to the length of the sampling time vector, and the position identifies the host (unless named
vectors are provided). Host names can be provided separately;
otherwise it will be tried to extract them from the sequences or sampling times.
It is also possible to include (otherwise unobserved) simulated data: sim.infection times, sim.infectors, 
and a (phylogenetic) sim.tree. This is done automatically when using \code{\link{sim.phybreak}}.
}
\examples{
sampletimedata <- c(0,2,2,4,4)
sampleSNPdata <- matrix(c("a","a","a","a","a",
                          "a","c","c","c","c",
                          "t","t","t","g","g"), nrow = 5, 
                          dimnames = list(LETTERS[1:5], NULL))
dataset <- phybreakdata(sequences = sampleSNPdata, sample.times = sampletimedata)
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
\references{
\href{http://dx.doi.org/10.1101/069195}{Klinkenberg et al, on biorXiv}.
}

