% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phybreak.R
\name{sample_phybreak}
\alias{sample_phybreak}
\alias{sample.phybreak}
\title{Sampling from a phybreak MCMC-chain.}
\usage{
sample_phybreak(x, nsample, thin = 1, edgewise = 0.8, keepphylo = 0,
  withinhost_only = 0, parameter_frequency = 1, status_interval = 10)

sample.phybreak(...)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{nsample}{The number of samples to take.}

\item{thin}{The thinning to use (values after every \code{thin}'th iteration will be included in the posterior). 
Each iteration does one update of all parameters and tree updates with each host as focal host once.}

\item{edgewise}{The proportion of tree updates with the edgewise protocol, which keeps the within-host minitree topology
intact when proposing a new infection time and infector, and then resamples the within-host minitree for each sampling tip
separately ("edgewise"). This results in better mixing of the mcmc-chain if there is much genetic information (many SNPs)
and if there are many possible within-host minitree topologies (e.g. many samples per host).}

\item{keepphylo}{The proportion of tree updates keeping the phylogenetic tree intact, only possible if there is one 
sample per host and the \code{wh.model = "linear"} with complete bottleneck.}

\item{withinhost_only}{The proportion of tree updates in which only the within-host minitree is sampled, and 
the transmission tree and infection times are kept unchanged.}

\item{parameter_frequency}{The relative frequency by which the model parameters are updated relative to updating each host.}

\item{status_interval}{The number of second between each on-screen print of the progress of the mcmc-chain.}
}
\value{
The \code{phybreak} object used to call the function, including (additional) samples from the posterior.
}
\description{
Function to take (additional) samples from the posterior distribution of a phylogenetic and transmission tree 
  (plus associated parameters), within a \code{phybreak} object (\code{sample.phybreak} is deprecated).
}
\examples{
#First create a phybreak-object
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(dataset = simulation)

MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 50, thin = 2)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
