% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phybreak.R
\name{sample_phybreak}
\alias{sample_phybreak}
\alias{sample.phybreak}
\title{Sampling from a phybreak MCMC-chain.}
\usage{
sample_phybreak(x, nsample, thin = 1, keepphylo = NULL,
  withinhost_only = 0)

sample.phybreak(...)
}
\arguments{
\item{x}{An object of class \code{phybreak}.}

\item{nsample}{The number of samples to take.}

\item{thin}{The thinning to use (values after every \code{thin}'th iteration will be included in the posterior). 
Each iteration does one update of all parameters and tree updates with each host as focal host once.}

\item{keepphylo}{The proportion of tree updates keeping the phylotree intact, only possible if there is one sample per host
and the \code{wh.model = "linear"}. If \code{NULL} (default), it is set to 0.2 only in that case, otherwise to 0.}

\item{withinhost_only}{The proportion of tree updates in which only the within-host minitree is sampled, and 
the transmission tree and infection times are kept unchanged.}
}
\value{
The \code{phybreak} object used to call the function, including (additional) samples from the posterior.
}
\description{
Function to take (additional) samples from the posterior distribution of a phylogenetic and transmission tree 
  (plus associated parameters), within a \code{phybreak} object (\code{sample.phybreak} is deprecated).
}
\examples{
#First create a phybreak-object
simulation <- sim_phybreak(obsize = 5)
MCMCstate <- phybreak(dataset = simulation)

MCMCstate <- burnin_phybreak(MCMCstate, ncycles = 20)
MCMCstate <- sample_phybreak(MCMCstate, nsample = 50, thin = 2)
}
\references{
\href{http://dx.doi.org/10.1371/journal.pcbi.1005495}{Klinkenberg et al. (2017)} Simultaneous 
  inference of phylogenetic and transmission trees in infectious disease outbreaks. 
  \emph{PLoS Comput Biol}, \strong{13}(5): e1005495.
}
\author{
Don Klinkenberg \email{don@xs4all.nl}
}
